Move previous data getter to core templatetags
