+++ payroll/views.py
@@ -10,6 +10,7 @@ from django.views.generic import CreateView, DeleteView, UpdateView
 
 from core.constants import MONTHS
 from core.models import FinancialYear
+from core.templatetags.util import get_previous_months_data
 from costcentre.models import CostCentre
 from payroll.forms import VacancyForm
 from payroll.models import Vacancy
@@ -107,7 +108,7 @@ class EditPayrollPage(EditPayrollBaseView):
         payroll_forecast_report_data = payroll_service.payroll_forecast_report(
             self.cost_centre, self.financial_year
         )
-        previous_months = list(payroll_service.get_previous_months_data())
+        previous_months = get_previous_months_data()
 
         context = {
             "cost_centre_code": self.cost_centre.cost_centre_code,
